communication: get rid of it
GazeboContext is just used by GazeboRealsense sensor and is mainly ref
couting the context which is for internal use only. So this patch move
this class to be part of the sensor implementation with no need to be
exported with the library.
