--- src/communication/GazeboContext.hh
@@ -1,32 +1,0 @@
-/*
-// Copyright (c) 2016 Intel Corporation
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-*/
-#pragma once
-
-#include <memory>
-#include <gazebo/gazebo_client.hh>
-#include <gazebo/transport/transport.hh>
-
-class GazeboContext
-{
-  public:
-    ~GazeboContext();
-    gazebo::transport::Node *node();
-    static std::shared_ptr<GazeboContext> instance();
-
-  private:
-    GazeboContext();
-};
-
